requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
langchain==0.1.0
langchain-openai==0.0.2
dspy-ai==2.4.0
openai==1.6.1
yt-dlp==2023.12.30
youtube-search-python==1.6.6
python-dotenv==1.0.0
httpx==0.25.2